+++ Changelog.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix write `route.json` file permission issue cause entire service corrupted in app.js. Completed on 2025-10-07
+- Fix out of control the create new mariadb classs in sqlmanager engine. Completed on 2025-10-08
 
 ### Security